src/main/java/com/shahed/taskmanagementsystem/utils/DateUtils.java
package com.shahed.taskmanagementsystem.utils;

import java.time.LocalDateTime;
import java.time.OffsetDateTime;
import java.time.ZoneOffset;
import lombok.NoArgsConstructor;

@NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
public final class DateUtils {
    public static LocalDateTime convertOffsetDtoToInternalDto(OffsetDateTime offsetDateTime) {
        if (offsetDateTime == null) {
            return null;
        }

        return offsetDateTime.atZoneSameInstant(ZoneOffset.UTC).toLocalDateTime();
    }
}
